refactor: structure code and file naming for importing data
