+++ backend/data_import/db/excluded_fields.py
@@ -1,4 +1,8 @@
-class APIResponseKeysToRemove:
+class SchoolFieldExclusions:
+    """
+    Fields from API response class that should be excluded during school data decomposition.
+    """
+
     TYP: str = "typ"
     ETAPY_EDUKACJI: str = "etapy_edukacji"
     MIEJSCOWOSC: str = "miejscowosc"
